+++ src/ui_bevy/ui_ext/text_rendering.rs
@@ -9,6 +9,7 @@ use super::{TextOutline, TextShadowGroup};
 //-------------------------------------------------------------------------------------------------------------------
 
 fn extract_text_outlines(
+    mut aa_glyph_cache: Local<Vec<ExtractedGlyph>>,
     mut commands: Commands,
     mut extracted_uinodes: ResMut<ExtractedUiNodes>,
     texture_atlases: Extract<Res<Assets<TextureAtlasLayout>>>,
@@ -27,6 +28,7 @@ fn extract_text_outlines(
     camera_map: Extract<UiCameraMap>,
 )
 {
+    aa_glyph_cache.clear();
     let mut start = extracted_uinodes.glyphs.len();
     let mut len = 0;
 
@@ -35,7 +37,7 @@ fn extract_text_outlines(
         &uinode_query
     {
         // Skip if not visible or if size is set to zero (e.g. when a parent is set to `Display::None`)
-        if !inherited_visibility.get() || uinode.is_empty() || outline.width < 1.0 {
+        if !inherited_visibility.get() || uinode.is_empty() || outline.width == 0.0 {
             continue;
         }
 
@@ -43,9 +45,12 @@ fn extract_text_outlines(
             continue;
         };
 
-        let width = outline.width as i32;
+        let width = (outline.width / uinode.inverse_scale_factor()).ceil() as i32;
+        let width_pow2 = width.pow(2);
         let aa_factor = outline.anti_aliasing.unwrap_or(1.0);
         let color: LinearRgba = outline.color.into();
+        let mut aa_color = color;
+        aa_color.alpha *= aa_factor;
 
         for (i, PositionedGlyph { position, atlas_info, .. }) in text_layout_info.glyphs.iter().enumerate() {
             let rect = texture_atlases
@@ -55,30 +60,30 @@ fn extract_text_outlines(
                 .as_rect();
 
             for offset_x in -width..=width {
-                for offset_y in -width..=width {
+                // Adjust height to follow a radial pattern.
+                let height = ((width_pow2 - offset_x.pow(2)).abs() as f32).sqrt().ceil() as i32;
+
+                for offset_y in -height..=height {
                     if offset_x == 0 && offset_y == 0 {
                         continue;
-                    }
-
-                    // Anti-aliasing.
-                    let mut color = color;
-                    if (offset_x.abs() == width) || (offset_y.abs() == width) {
-                        color.alpha *= aa_factor;
                     }
 
                     let offset = Vec2 { x: offset_x as f32, y: offset_y as f32 };
 
                     let transform = global_transform.affine()
-                        * Mat4::from_translation(
-                            (-0.5 * uinode.size() + offset / uinode.inverse_scale_factor()).extend(0.),
-                        );
+                        * Mat4::from_translation((-0.5 * uinode.size() + offset).extend(0.));
 
-                    extracted_uinodes.glyphs.push(ExtractedGlyph {
+                    let extracted_glyph = ExtractedGlyph {
                         transform: transform * Mat4::from_translation(position.extend(0.)),
                         rect,
-                    });
+                    };
 
-                    len += 1;
+                    if aa_factor != 1.0 && offset_y.abs() == height {
+                        aa_glyph_cache.push(extracted_glyph);
+                    } else {
+                        extracted_uinodes.glyphs.push(extracted_glyph);
+                        len += 1;
+                    }
                 } // y offset
             } // x offset
 
@@ -87,6 +92,11 @@ fn extract_text_outlines(
                 .get(i + 1)
                 .is_none_or(|info| info.atlas_info.texture != atlas_info.texture)
             {
+                let aa_len = aa_glyph_cache.len();
+                for aa_glyph in aa_glyph_cache.drain(..) {
+                    extracted_uinodes.glyphs.push(aa_glyph);
+                }
+
                 extracted_uinodes.uinodes.push(ExtractedUiNode {
                     render_entity: commands.spawn(TemporaryRenderEntity).id(),
                     stack_index: uinode.stack_index,
@@ -100,6 +110,20 @@ fn extract_text_outlines(
                 });
                 start += len;
                 len = 0;
+
+                if aa_len > 0 {
+                    extracted_uinodes.uinodes.push(ExtractedUiNode {
+                        render_entity: commands.spawn(TemporaryRenderEntity).id(),
+                        stack_index: uinode.stack_index,
+                        color: aa_color,
+                        image: atlas_info.texture.id(),
+                        clip: clip.map(|clip| clip.clip),
+                        extracted_camera_entity,
+                        rect,
+                        item: ExtractedUiItem::Glyphs { range: start..(start + aa_len) },
+                        main_entity: entity.into(),
+                    });
+                }
             }
         }
     }